chore(web): components relocate
